Struct automata::Stack

pub struct Stack {
    pub delta: Delta,
}

A stack of labelled transitions

Fields
delta: Delta
Implementations

impl Stack

pub fn new() -> Stack

pub fn expr_0(&mut self) -> Result<AutomataRef, &'static str>

Returns reference to an acceptor that never transitions to a final state, i.e. accept nothing

Examples

Example 1:

use automata::{ ANFA, AutomataRef, Stack };

let mut machine_stack = Stack::new();
// it's always safe to unwrap acceptor automata
let ref_0: AutomataRef = machine_stack.expr_0().unwrap();
let machine: ANFA = ANFA::from((ref_0, machine_stack));

Implementation

AutomataRef

AutomataRef {
    q0: 0,
    f: 1,
}

Stack

Stack {
    delta: [
        [],
        []
    ]
}

Into ANFA

{
    delta: [
        [],
        []
    ],
    q0: 0,
    f: 1,
}

Definition

(
  Q: { 0, 1 },
  δ: none,
  q0: 0,
  F: 1
)

 ( 0 )  (( 1 ))

pub fn expr_1(&mut self) -> Result<AutomataRef, &'static str>

Returns reference to an acceptor in final state, i.e. accept anything, AKA epsilon acceptor

Examples

Example 1:

use automata::{ ANFA, AutomataRef, Stack };

let mut machine_stack = Stack::new();
// it's always safe to unwrap acceptor automata
let ref_0: AutomataRef = machine_stack.expr_1().unwrap();
let machine: ANFA = ANFA::from((ref_0, machine_stack));

Implementation

AutomataRef

AutomataRef {
    q0: 0,
    f: 0,
}

Stack

Stack {
    delta: [
        []
    ],
}

Into ANFA

ANFA {
    delta: [
        []
    ],
    q0: 0,
    f: 0
}

Definition

(
  Q: { 0 },
  δ: none,
  q0: 0,
  F: 0
)

 (( 0 ))

pub fn expr_a(&mut self, c: char) -> Result<AutomataRef, &'static str>

Returns reference to an automaton accepting the provided any in Σ

Examples

Example 1:

use automata::{ ANFA, AutomataRef, Stack };

let mut machine_stack = Stack::new();
// it's always safe to unwrap acceptor automata
let ref_0: AutomataRef = machine_stack.expr_a('a').unwrap();
let machine: ANFA = ANFA::from((ref_0, machine_stack));

Implementation

AutomataRef

AutomataRef {
    q0: 0,
    f: 1
}

Stack

Stack {
    delta: [
        [(Some('a'), 1)],
        []
    ]
}

Into ANFA

ANFA {
    delta: [
        [(Some('a'), 1)],
        []
    ],
    q0: 0,
    f: 1
}

Definition

(
  Q: { 0, 1 },
  δ: (0, 'a') => 1,
  q0: 0,
  F: 1
)

 machine a
 ( 0 ) -- 'a' --> (( 1 ))

pub fn concatenate(
&mut self,
machine_ref_a: &AutomataRef,
machine_ref_b: &AutomataRef
) -> Result<AutomataRef, &'static str>

Concatenates machine slices a and b of the same stack

Concatenation is an associative, binary operation:

machine_c = machine_a ⋅ machine_b
machine_n = machine_a ⋅ machine_b ⋅ machine_c
machine_n = (machine_a ⋅ machine_b) ⋅ machine_c
machine_n = machine_a ⋅ (machine_b ⋅ machine_c)

Examples

Example 1:

use automata::{ ANFA, AutomataRef, Stack };

let mut machine_stack = Stack::new();
let slice_a: AutomataRef = machine_stack.expr_a('a').unwrap();
let slice_b: AutomataRef = machine_stack.expr_a('b').unwrap();
let slice_c_result: Result<AutomataRef, &'static str> = machine_stack.concatenate(&slice_a, &slice_b);
match slice_c_result {
    Ok(slice_c) => {
        println!("{:#?}", ANFA::from((slice_c, machine_stack)));
    },
    Err(err) => {
        println!("Error creating automaton: {}", err);
    }
};

Diagram

 machine_a
 ( 0 ) --> 'a' --> (( 1 ))

 machine_b
 ( 2 ) --> 'b' --> (( 3 ))

 machine_c
 ( 0 ) -- 'a' --> ( 1 ) -- ε --> ( 2 ) -- 'b' --> (( 3 ))

pub fn star(
&mut self,
machine_ref_a: &AutomataRef
) -> Result<AutomataRef, &'static str>

From slice a, pushes the star operation – a* – to the stack, returning slice b

Star is a unary operation:

machine_b = machine_a*

Examples

Diagram

machine_a
( 0 ) -- 'a' --> (( 1 ))

machine b = machine_a*
                    /-- 0 --> ( 0 ) -- 'a' --> ( 1 )
( 2 ) -- ε --> ( 3 ) <------------ ε ------------|
                    \-- 1 --> (( 4 ))

pub fn union(
&mut self,
machine_ref_a: &AutomataRef,
machine_ref_b: &AutomataRef
) -> Result<AutomataRef, &'static str>

Diagram

machine_a
( 0 ) -- 'a' --> (( 1 ))

machine_b
( 2 ) -- 'b' --> (( 3 ))

machine_c = machina_a ∪ machine_b
    / -- 0 --> ( 0 ) -- 'a' --> ( 1 ) --\
( 4 )                                    ε --> (( 5 ))
    \ -- 1 --> ( 2 ) -- 'b' --> ( 2 ) --/